requirements.txt
cffi==1.11.5
django>=2.1.2
django-bootstrap3==7.0.1
django-debug-toolbar==1.5
misaka==2.0.0
sqlparse==0.2.1
asn1crypto==0.24.0
cryptography==2.4.1
enum34==1.1.6
idna==2.7
ipaddress==1.0.22
pycparser==2.19
pymysql==0.9.2
six==1.11.0
mysqlclient==1.3.13
locustio==0.9.0
psycopg2